community detailed view

+++ GamesHub/GamesBuzz/models.py
@@ -18,7 +18,7 @@ class GameInteraction(models.Model):
     rating         = models.PositiveSmallIntegerField(default= 0, choices=RATING_CHOICES, null=True, blank=True) 
     purchase_date  = models.DateTimeField(auto_now_add=True)
     purchase_price = models.FloatField(default=0)
-    transaction_id = models.IntegerField(primary_key=True)
+    transaction_id = models.BigAutoField(primary_key=True)
     in_library     = models.BooleanField(default=True)
 
     class Meta:
